+++ front/src/components/SettingsPanel/SettingsContent.jsx
@@ -403,11 +403,6 @@ const SettingsPanel = ({ localState, setLocalState, userData, modelsData }) => {
           {/* Warning for external models */}
           <DataSafetyText localState={localState} userData={userData} />
           {/* Use Tools – checkbox */}
-          {/* <ToolsToggle localState={localState} setLocalState={setLocalState} />
-          <WebSearchToggle
-            localState={localState}
-            setLocalState={setLocalState}
-          /> */}
           {toolsModule && <ToolsContainer
               localState={localState}
               setLocalState={setLocalState}