+++ front/src/modals/ModalContext.jsx
@@ -25,7 +25,6 @@ import DeleteFolderModal from "../components/Sidebar/DeleteFolderModal";
 import MoveConversationModal from "../components/Sidebar/MoveConversationModal";
 import ClearMemoryModal from "./UserSettings/ClearMemoryModal";
 import ClearMessagesModal from "./Alert/ClearMessagesModal";
-import SummarizeReplaceModal from "./Alert/SummarizeReplaceModal";
 import PreviewModal from "./Chat/PreviewModal";
 import UnsentFilesModal from "./Alert/UnsentFilesModal";
 import UnprocessedFilesModal from "./Alert/UnprocessedFilesModal";
@@ -157,9 +156,6 @@ export function ModalProvider({ children }) {
       )}
       {modalType === "clearMessages" && (
         <ClearMessagesModal isOpen onClose={closeModal} {...modalProps} />
-      )}
-      {modalType === "summarizeReplace" && (
-        <SummarizeReplaceModal isOpen onClose={closeModal} {...modalProps} />
       )}
       {/* General Modals */}
       {modalType === "exportConversation" && (